refactor: extract wiring helpers from main.go into helpers.go
